ESP8266: SDK and LwIP update
SDK is updated to https://github.com/espressif/ESP8266_NONOS_SDK/commit/a0b131126ba803d0069014698a07e9f2fd3decd6
Tracking branch: https://github.com/espressif/ESP8266_NONOS_SDK/compare/release/v2.2.x

LwIP is updated to 2.1.2.

+++ src/mgos_debug.c
@@ -90,7 +90,7 @@ void mgos_debug_write(int fd, const void *data, size_t len) {
         buf, sizeof(buf), "%s %u %.3lf %d|",
         (mgos_sys_config_get_device_id() ? mgos_sys_config_get_device_id()
                                          : "-"),
-        s_seq, mg_time(), fd);
+        (unsigned int) s_seq, mg_time(), fd);
     if (n > 0) {
       mgos_debug_udp_send(mg_mk_str_n(buf, n), mg_mk_str_n(data, len));
     }
